requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
crawl4ai>=0.3.0
google-cloud-storage>=2.10.0
python-multipart>=0.0.6 